plugins/Kingdom-Scheduler/pyproject.toml
[tool.poetry]
name = "kingdom-scheduler"
version = "0.1.0"
description = ""
authors = ["Two Faced Janus. <blcklptn@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.7.3"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
motor = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
